[WIP] load providers from config and resolve link to CWL package
