+++ .gitignore
@@ -10,6 +10,7 @@
 /data_sources.json
 /custom_wps.cfg
 /Makefile.config
+/wps_providers.yml
 
 # Python / Extensions etc.
 *~
