+++ buildout.cfg
@@ -66,6 +66,7 @@ wps-restapi-ref =
 wps-email-notify-from-addr =
 wps-email-notify-password =
 wps-email-notify-template =
+wps-providers-file =
 extra-options =
 data-sources =
