+++ templates/weaver.ini
@@ -35,6 +35,7 @@ weaver.wps_restapi_ref = ${settings:wps-restapi-ref}
 weaver.wps_email_notify_from_addr = ${settings:wps-email-notify-from-addr}
 weaver.wps_email_notify_password = ${settings:wps-email-notify-password}
 weaver.wps_email_notify_template = ${settings:wps-email-notify-template}
+weaver.wps_providers_file = ${settings:wps-providers-file}
 weaver.data_sources = ${settings:data-sources}
 weaver.extra_options = ${settings:extra-options}
