+++ weaver/processes/utils.py
@@ -1,11 +1,47 @@
+from weaver.config import get_weaver_configuration, WEAVER_CONFIGURATION_EMS
+from weaver.datatype import Service, Process as ProcessDB
+from weaver.database import get_db
+from weaver.store.base import StoreProcesses
+from weaver.utils import get_sane_name, get_settings
+from weaver.processes import wps_package
+from weaver.processes.types import PROCESS_APPLICATION, PROCESS_WORKFLOW
+from weaver.wps_restapi import swagger_definitions as sd
+from weaver.wps_restapi.utils import wps_restapi_base_url
+from weaver.exceptions import (
+    InvalidIdentifierValue,
+    ProcessRegistrationError,
+    PackageRegistrationError,
+    PackageTypeError,
+    PackageNotFound,
+)
 from owslib.wps import ComplexData, is_reference
+from owslib.wps import WebProcessingService
+from pyramid.httpexceptions import (
+    HTTPOk,
+    HTTPNotFound,
+    HTTPBadRequest,
+    HTTPConflict,
+    HTTPUnprocessableEntity,
+    HTTPInternalServerError,
+    HTTPException,
+)
 from six.moves.urllib.request import urlopen
+from six.moves.urllib.parse import urlparse, urlunsplit, parse_qs
 from six.moves.urllib.error import URLError
 from typing import TYPE_CHECKING
+import colander
+import warnings
+import logging
+import yaml
 import json
+import six
+import os
 if TYPE_CHECKING:
-    from weaver.typedefs import JsonBody
+    from weaver.typedefs import JsonBody, AnyContainer, AnySettingsContainer, FileSystemPathType
+    from typing import AnyStr, Dict, Union
+    from pywps import Process as ProcessWPS
     import owslib.wps
+LOGGER = logging.getLogger(__name__)
 
 
 def _get_data(input_value):
@@ -103,3 +139,228 @@ def jsonify_value(value):
     # other type
     else:
         return value
+
+
+def convert_process_wps_to_db(service, process, container):
+    # type: (Union[Service, Dict[{'url': AnyStr, 'name': AnyStr}]], ProcessWPS, AnySettingsContainer) -> ProcessDB
+    """
+    Converts an owslib WPS Process to local storage Process.
+    """
+    describe_process_url = '{base_url}/providers/{provider_id}/processes/{process_id}'.format(
+        base_url=wps_restapi_base_url(container),
+        provider_id=service.get('name'),
+        process_id=process.identifier)
+    execute_process_url = '{describe_url}/jobs'.format(describe_url=describe_process_url)
+
+    default_format = {'mimeType': 'text/plain'}
+    inputs = [dict(
+        id=getattr(dataInput, 'identifier', ''),
+        title=getattr(dataInput, 'title', ''),
+        abstract=getattr(dataInput, 'abstract', ''),
+        minOccurs=str(getattr(dataInput, 'minOccurs', 0)),  # FIXME: str applied to match OGC REST-API definition
+        maxOccurs=str(getattr(dataInput, 'maxOccurs', 0)),  # FIXME: str applied to match OGC REST-API definition
+        dataType=dataInput.dataType,
+        defaultValue=jsonify_value(getattr(dataInput, 'defaultValue', None)),
+        allowedValues=[jsonify_value(dataValue) for dataValue in getattr(dataInput, 'allowedValues', [])],
+        supportedValues=[jsonify_value(dataValue) for dataValue in getattr(dataInput, 'supportedValues', [])],
+        formats=[jsonify_value(dataValue) for dataValue in getattr(dataInput, 'supportedValues', [default_format])],
+    ) for dataInput in getattr(process, 'dataInputs', [])]
+
+    outputs = [dict(
+        id=getattr(processOutput, 'identifier', ''),
+        title=getattr(processOutput, 'title', ''),
+        abstract=getattr(processOutput, 'abstract', ''),
+        dataType=processOutput.dataType,
+        defaultValue=jsonify_value(getattr(processOutput, 'defaultValue', None)),
+        formats=[jsonify_value(dataValue) for dataValue in getattr(processOutput, 'supportedValues', [default_format])],
+    ) for processOutput in getattr(process, 'processOutputs', [])]
+
+    return ProcessDB(
+        id=process.identifier,
+        label=getattr(process, 'title', ''),
+        title=getattr(process, 'title', ''),
+        abstract=getattr(process, 'abstract', ''),
+        inputs=inputs,
+        outputs=outputs,
+        url=describe_process_url,
+        processEndpointWPS1=service.get('url'),
+        processDescriptionURL=describe_process_url,
+        executeEndpoint=execute_process_url,
+        package=None,
+    )
+
+
+def convert_process_ems_to_db():
+    pass
+
+
+def add_process_from_payload(payload, container):
+    # type: (JsonBody, AnyContainer) -> HTTPException
+    """
+    Adds a :class:`weaver.datatype.Process` instance to storage using the provided JSON ``payload`` matching
+    :class:`weaver.wps_restapi.swagger_definitions.ProcessDescription`.
+    """
+
+    # validate minimum field requirements
+    try:
+        sd.Deploy().deserialize(payload)
+    except colander.Invalid as ex:
+        raise HTTPBadRequest("Invalid schema: [{}]".format(str(ex)))
+    except Exception as ex:
+        raise HTTPInternalServerError("Unhandled error when parsing 'processDescription': [{}]".format(str(ex)))
+
+    # validate identifier naming for unsupported characters
+    process_description = payload.get('processDescription')
+    process_info = process_description.get('process')
+    process_href = process_description.pop('href', None)
+
+    # retrieve CWL package definition, either via 'href' (WPS-1/2), 'owsContext' or 'executionUnit' (package/reference)
+    deployment_profile_name = payload.get('deploymentProfileName', '').lower()
+    ows_context = process_info.pop('owsContext', None)
+    reference = None
+    package = None
+    if process_href:
+        reference = process_href  # reference type handled downstream
+    elif isinstance(ows_context, dict):
+        offering = ows_context.get('offering')
+        if not isinstance(offering, dict):
+            raise HTTPUnprocessableEntity("Invalid parameter 'processDescription.process.owsContext.offering'.")
+        content = offering.get('content')
+        if not isinstance(content, dict):
+            raise HTTPUnprocessableEntity("Invalid parameter 'processDescription.process.owsContext.offering.content'.")
+        package = None
+        reference = content.get('href')
+    elif deployment_profile_name:
+        if not any(deployment_profile_name.endswith(typ) for typ in [PROCESS_APPLICATION, PROCESS_WORKFLOW]):
+            raise HTTPBadRequest("Invalid value for parameter 'deploymentProfileName'.")
+        execution_units = payload.get('executionUnit')
+        if not isinstance(execution_units, list):
+            raise HTTPUnprocessableEntity("Invalid parameter 'executionUnit'.")
+        for execution_unit in execution_units:
+            if not isinstance(execution_unit, dict):
+                raise HTTPUnprocessableEntity("Invalid parameter 'executionUnit'.")
+            package = execution_unit.get('unit')
+            reference = execution_unit.get('href')
+            # stop on first package/reference found, simultaneous usage will raise during package retrieval
+            if package or reference:
+                break
+    else:
+        raise HTTPBadRequest("Missing one of required parameters [owsContext, deploymentProfileName].")
+
+    # obtain updated process information using WPS process offering, CWL/WPS reference or CWL package definition
+    try:
+        # data_source `None` forces workflow process to search locally for deployed step applications
+        process_info = wps_package.get_process_definition(process_info, reference, package, data_source=None)
+    except PackageNotFound as ex:
+        # raised when a workflow sub-process is not found (not deployed locally)
+        raise HTTPNotFound(detail=str(ex))
+    except (PackageRegistrationError, PackageTypeError) as ex:
+        raise HTTPUnprocessableEntity(detail=str(ex))
+    except InvalidIdentifierValue as ex:
+        raise HTTPBadRequest(str(ex))
+    except Exception as ex:
+        raise HTTPBadRequest("Invalid package/reference definition. Loading generated error: [{!r}]".format(ex))
+
+    # validate process type against weaver configuration
+    settings = get_settings(container)
+    process_type = process_info['type']
+    if process_type == PROCESS_WORKFLOW:
+        weaver_config = get_weaver_configuration(settings)
+        if weaver_config != WEAVER_CONFIGURATION_EMS:
+            raise HTTPBadRequest("Invalid [{0}] package deployment on [{1}].".format(process_type, weaver_config))
+
+    restapi_url = wps_restapi_base_url(settings)
+    description_url = "/".join([restapi_url, 'processes', process_info['identifier']])
+    execute_endpoint = "/".join([description_url, "jobs"])
+
+    # ensure that required 'processEndpointWPS1' in db is added,
+    # will be auto-fixed to localhost if not specified in body
+    process_info['processEndpointWPS1'] = process_description.get('processEndpointWPS1')
+    process_info['executeEndpoint'] = execute_endpoint
+    process_info['payload'] = payload
+    process_info['jobControlOptions'] = process_description.get('jobControlOptions', [])
+    process_info['outputTransmission'] = process_description.get('outputTransmission', [])
+    process_info['processDescriptionURL'] = description_url
+    # insert the "resolved" context using details retrieved from 'executionUnit'/'href' or directly with 'owsContext'
+    if 'owsContext' not in process_info and reference:
+        process_info['owsContext'] = {'offering': {'content': {'href': reference}}}
+    elif isinstance(ows_context, dict):
+        process_info['owsContext'] = ows_context
+
+    try:
+        store = get_db(container).get_store(StoreProcesses)
+        saved_process = store.save_process(ProcessDB(process_info), overwrite=False)
+    except ProcessRegistrationError as ex:
+        raise HTTPConflict(detail=str(ex))
+    except ValueError as ex:
+        # raised on invalid process name
+        raise HTTPBadRequest(detail=str(ex))
+
+    json_response = {'processSummary': saved_process.process_summary(), 'deploymentDone': True}
+    return HTTPOk(json=json_response)
+
+
+def register_wps_provider_processes(wps_providers_file_path, container):
+    # type: (FileSystemPathType, AnySettingsContainer) -> None
+    """
+    Loads a `wps_provider.yml` file and registers WPS-1 providers processes to the
+    current `Weaver` instance as equivalent WPS-2 processes.
+
+    .. seealso::
+        - `weaver.wps_providers.yml.example`
+    """
+    if not os.path.isfile(wps_providers_file_path):
+        warnings.warn("No file specified for WPS providers registration.", RuntimeWarning)
+        return
+    try:
+        with open(wps_providers_file_path, 'r') as f:
+            providers_config = yaml.safe_load(f)
+        providers = providers_config['providers']
+        from weaver.wps_restapi.processes.processes import list_remote_processes
+        for cfg_service in providers:
+            # parse info
+            if isinstance(cfg_service, dict):
+                svc_url = cfg_service['url']
+                svc_name = cfg_service.get('name')
+                svc_proc = cfg_service.get('processes', [])
+            elif isinstance(cfg_service, six.string_types):
+                svc_url = cfg_service
+                svc_name = None
+                svc_proc = []
+            else:
+                raise ValueError("Invalid service value: [{!s}].".format(cfg_service))
+            url_p = urlparse(svc_url)
+            qs_p = parse_qs(url_p.query)
+            svc_url = urlunsplit(url_p[:4] + tuple(['']))  # remove any query in url
+            svc_name = svc_name or get_sane_name(url_p.hostname)
+            svc_proc = svc_proc or qs_p.get('identifier', [])
+            if not isinstance(svc_name, six.string_types):
+                raise ValueError("Invalid service value: [{!s}].".format(svc_name))
+            if not isinstance(svc_proc, list):
+                raise ValueError("Invalid process value: [{!s}].".format(svc_proc))
+
+            # fetch data
+            LOGGER.info("Fetching WPS-1: [{}]".format(svc_url))
+            wps = WebProcessingService(url=svc_url)
+            wps_processes = [wps.describeprocess(p) for p in svc_proc] or wps.processes
+            for wps_process in wps_processes:
+                proc_id = '{}_{}'.format(svc_name, get_sane_name(wps_process.identifier))
+                payload = {
+                    'processDescription': {
+                        'process': {
+                            'identifier': proc_id,
+                            'owsContext': {'offering': {'content': {'href': '{}'.format(svc_url)}}},
+                        }
+                    }
+                }
+                try:
+                    add_process_from_payload(payload, container)
+                except HTTPConflict:
+                    LOGGER.warning("Process already registered: [{}]".format(proc_id))
+                    continue
+                LOGGER.info("Process registered: []".format(proc_id))
+
+    except Exception as exc:
+        msg = "Invalid WPS providers configuration file [{!r}].".format(exc)
+        LOGGER.exception(msg)
+        raise RuntimeError(msg)
