+++ weaver/wps_restapi/api.py
@@ -1,8 +1,8 @@
 from weaver.__meta__ import __version__ as weaver_version
-from weaver.utils import get_settings
+from weaver.utils import get_settings, get_header
 from weaver.wps_restapi import swagger_definitions as sd
 from weaver.wps_restapi.colander_one_of import CustomTypeConversionDispatcher
-from weaver.wps_restapi.utils import wps_restapi_base_url, wps_restapi_base_path, get_header, CONTENT_TYPE_JSON
+from weaver.wps_restapi.utils import wps_restapi_base_url, wps_restapi_base_path, CONTENT_TYPE_JSON
 from weaver.owsexceptions import OWSException
 from six.moves.urllib.parse import urlparse
 from cornice_swagger import CorniceSwagger
